feat(host): Export default CLIs
