+++ bin/host/src/interop/orchestrator.rs
@@ -1,15 +1,14 @@
 //! [SingleChainHostCli]'s [HostOrchestrator] + [DetachedHostOrchestrator] implementations.
 
 use super::{InteropFetcher, InteropHostCli, LocalKeyValueStore};
-use crate::eth::http_provider;
+use crate::{
+    eth::http_provider, DetachedHostOrchestrator, DiskKeyValueStore, Fetcher, HostOrchestrator,
+    MemoryKeyValueStore, SharedKeyValueStore, SplitKeyValueStore,
+};
 use alloy_primitives::map::HashMap;
 use alloy_provider::{Provider, RootProvider};
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
-use kona_host::{
-    DetachedHostOrchestrator, DiskKeyValueStore, Fetcher, HostOrchestrator, MemoryKeyValueStore,
-    SharedKeyValueStore, SplitKeyValueStore,
-};
 use kona_preimage::{HintWriter, NativeChannel, OracleReader};
 use kona_providers_alloy::{OnlineBeaconClient, OnlineBlobProvider};
 use std::sync::Arc;
