+++ bin/host/src/single/local_kv.rs
@@ -2,9 +2,9 @@
 //! using the [SingleChainHostCli] config.
 
 use super::SingleChainHostCli;
+use crate::KeyValueStore;
 use alloy_primitives::B256;
 use anyhow::Result;
-use kona_host::KeyValueStore;
 use kona_preimage::PreimageKey;
 use kona_proof::boot::{
     L1_HEAD_KEY, L2_CHAIN_ID_KEY, L2_CLAIM_BLOCK_NUMBER_KEY, L2_CLAIM_KEY, L2_OUTPUT_ROOT_KEY,
