+++ bin/host/src/single/orchestrator.rs
@@ -1,14 +1,13 @@
 //! [SingleChainHostCli]'s [HostOrchestrator] + [DetachedHostOrchestrator] implementations.
 
 use super::{LocalKeyValueStore, SingleChainFetcher, SingleChainHostCli};
-use crate::eth::http_provider;
-use alloy_provider::RootProvider;
+use crate::{
+    eth::http_provider, DetachedHostOrchestrator, DiskKeyValueStore, Fetcher, HostOrchestrator,
+    MemoryKeyValueStore, SharedKeyValueStore, SplitKeyValueStore,
+};
+use alloy_provider::ReqwestProvider;
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
-use kona_host::{
-    DetachedHostOrchestrator, DiskKeyValueStore, Fetcher, HostOrchestrator, MemoryKeyValueStore,
-    SharedKeyValueStore, SplitKeyValueStore,
-};
 use kona_preimage::{HintWriter, NativeChannel, OracleReader};
 use kona_providers_alloy::{OnlineBeaconClient, OnlineBlobProvider};
 use std::sync::Arc;
